refactor: 리뷰 반영
CDN 코드 박스 안에 모두 들어가도록 수정
